+++ src/library/DIPS.Mobile.UI/Components/Chips/dotnet/ChipHandler.cs
@@ -5,7 +5,24 @@ namespace DIPS.Mobile.UI.Components.Chips;
 
 public partial class ChipHandler : ViewHandler<Chip, ViewHandler>
 {
-    private static partial void MapTitle(ChipHandler handler, Chip chip) { }
+    private static partial void MapTitle(ChipHandler handler, Chip chip) => throw new Only_Here_For_UnitTests();
+
+    private static partial void MapHasCloseButton(ChipHandler handler, Chip chip) =>
+        throw new Only_Here_For_UnitTests();
 
     protected override ViewHandler CreatePlatformView() => throw new Only_Here_For_UnitTests();
+    private static partial void MapColor(ChipHandler handler, Chip chip)
+    {
+        throw new Only_Here_For_UnitTests();
+    }
+
+    private static partial void MapCloseButtonColor(ChipHandler handler, Chip chip)
+    {
+        throw new Only_Here_For_UnitTests();
+    }
+
+    private static partial void MapCornerRadius(ChipHandler handler, Chip arg2)
+    {
+        throw new Only_Here_For_UnitTests();
+    }
 }